@@ -4,15 +4,15 @@
     "type": "project",
     "license": "MIT",
     "require": {
-        "php": ">=7.2.5",
+        "php": ">=8.0.2",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "symfony/console": "5.*",
-        "symfony/dotenv": "5.*",
+        "symfony/console": "6.*",
+        "symfony/dotenv": "6.*",
         "symfony/flex": "^1.17",
-        "symfony/framework-bundle": "5.*",
-        "symfony/runtime": "*",
-        "symfony/yaml": "5.*"
+        "symfony/framework-bundle": "6.*",
+        "symfony/runtime": "6.*",
+        "symfony/yaml": "6.*"
     },
     "require-dev": {
         "yceruto/ddd-maker-bundle": "^1.1"
@@ -66,7 +66,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "5.4.*"
+            "require": "6.0.*"
         }
     }
 }
